requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database adapters
psycopg[binary]==3.1.18
psycopg2-binary==2.9.7

# Task Queue
celery==5.3.4
redis==5.0.1

# Authentication & Firebase
firebase-admin==6.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Configuration & Deployment
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0

# CORS handling for frontend integration
django-cors-headers==4.3.1

# Additional utilities
Pillow==10.1.0

# Development dependencies
django-debug-toolbar==4.2.0

# Note: For GeoDjango on Windows, install GDAL via conda:
# conda install -c conda-forge gdal geos proj
